✨APIの実装を修正しまいた。
